did_you_mean without dependency
