+++ crates/nu-cli/Cargo.toml
@@ -55,7 +55,6 @@ indexmap = {version = "1.6.0", features = ["serde-1"]}
 itertools = "0.9.0"
 log = "0.4.11"
 meval = "0.2.0"
-natural = "0.5.0"
 num-bigint = {version = "0.3.0", features = ["serde"]}
 num-format = {version = "0.4.0", features = ["with-num-bigint"]}
 num-traits = "0.2.12"
