+++ crates/nu-protocol/Cargo.toml
@@ -20,7 +20,6 @@ getset = "0.1.1"
 indexmap = {version = "1.6.0", features = ["serde-1"]}
 itertools = "0.9.0"
 log = "0.4.11"
-natural = "0.5.0"
 nu-errors = {path = "../nu-errors", version = "0.20.0"}
 nu-source = {path = "../nu-source", version = "0.20.0"}
 num-bigint = {version = "0.3.0", features = ["serde"]}
