+++ crates/nu-protocol/src/lib.rs
@@ -18,8 +18,9 @@ pub use crate::signature::{NamedType, PositionalType, Signature};
 pub use crate::syntax_shape::SyntaxShape;
 pub use crate::type_name::{PrettyType, ShellTypeName, SpannedTypeName};
 pub use crate::type_shape::{Row as RowType, Type};
-pub use crate::value::column_path::{did_you_mean, ColumnPath, PathMember, UnspannedPathMember};
+pub use crate::value::column_path::{ColumnPath, PathMember, UnspannedPathMember};
 pub use crate::value::dict::{Dictionary, TaggedDictBuilder};
+pub use crate::value::did_you_mean::did_you_mean;
 pub use crate::value::evaluate::Scope;
 pub use crate::value::primitive::Primitive;
 pub use crate::value::primitive::{format_date, format_duration, format_primitive};
